requirements.txt
requests
bs4
psycopg2-binary